Massive changes. Most ui changes are related to the const const error or editing using an invalid version of qt. NOTE that this branch does not compile under qt 3.1.2 and the ui files were intentionally saved with qt 3.3.x. This may force the upgrade to qt 3.3.x prior to release of 0.7. Since 3.3.x is backward compatible with 3.1.2 there should be little impact (famous last words).
Projection selector uses sqlite3 database to store and fetch projection
information. Sqlite 3.x (latest) is now a required dependency to build
the projection branch. See http://www.sqlite.org.

Makefile.am files are hacked to include the sqlite3 library for linking.
build system needs to be tweaked to detect both sqlite3 and proj.

Projection detection and function still needs worked. This commit just lays
the groundwork for the new implementation.


git-svn-id: http://svn.osgeo.org/qgis@3056 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ branches/Projections_Branch/qgis/src/qgsmapcanvas.cpp
@@ -1463,7 +1463,7 @@ void QgsMapCanvas::mouseReleaseEvent(QMouseEvent * e)
       QgsPoint lastpoint = mCanvasProperties->coordXForm->transform(it->x(),it->y());
       QgsPoint endpoint = mCanvasProperties->coordXForm->transform(digitisedpoint.x(),digitisedpoint.y());
       paint.drawLine(static_cast<int>(lastpoint.x()),static_cast<int>(lastpoint.y()),
-         endpoint.x(),endpoint.y());
+         static_cast<int>(endpoint.x()),static_cast<int>(endpoint.y()));
       //draw it to an acetate layer
       QgsLine digitline(*it,digitisedpoint);
       QgsAcetateLines* acetate=new QgsAcetateLines();
@@ -1569,7 +1569,7 @@ void QgsMapCanvas::mouseReleaseEvent(QMouseEvent * e)
         QgsPoint point = mCanvasProperties->coordXForm->toMapCoordinates(e->x(), e->y());
 
   if ( !mMeasure ) {
-      mMeasure = new QgsMeasure(this, topLevelWidget() );
+    mMeasure = new QgsMeasure(this, topLevelWidget() );
   }
   mMeasure->addPoint(point);
   mMeasure->show();
